requirements.txt
albucore==0.0.24
albumentations==2.0.7
annotated-types==0.7.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
h11==0.16.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.8.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
